Move WeatherService into TestServices folder
